refactor(as/challenge): introduce Challenger trait + Fs/Ephemeral impls
Split the challenge.rs monolith into a challenge/ module: mod.rs holds the
async Challenger trait and the shared build_challenge_json/verify_jwt/rs384_sign
helpers, ephemeral.rs holds an in-memory-key EphemeralJwtChallenger, and fs.rs
(feature-gated) holds the on-disk-key FsJwtChallenger. AttestationService gains
a challenger field wired in new(), the challenge_key_path() accessor becomes
challenger(), and the grpc/restful bin handlers verify via challenger() instead
of the old free function.
