+++ attestation-service/src/bin/grpc/mod.rs
@@ -1,5 +1,5 @@
 use anyhow::bail;
-use attestation_service::challenge::verify_challenge_and_extract_nonce_b64url;
+use attestation_service::Challenger;
 use attestation_service::HashAlgorithm;
 use attestation_service::{
     config::Config, config::ConfigError, AttestationService as Service, ServiceError, Tee,
@@ -177,9 +177,12 @@ impl AttestationService for Arc<RwLock<AttestationServer>> {
                             .map_err(|e| Status::aborted(format!(
                                 "parse structured runtime data: {e}")))?;
                         if let Some(jwt) = structured.get("challenge_token").and_then(|x| x.as_str()) {
-                            let challenge_key_path =
-                                self.read().await.attestation_service.challenge_key_path();
-                            verify_challenge_and_extract_nonce_b64url(jwt, &challenge_key_path)
+                            self.read()
+                                .await
+                                .attestation_service
+                                .challenger()
+                                .verify_challenge_and_extract_nonce_b64url(jwt)
+                                .await
                                 .map_err(|e| Status::aborted(format!(
                                     "verify challenge_token failed: {e}")))?;
                         }
