+++ attestation-service/src/lib.rs
@@ -12,6 +12,7 @@ pub mod token;
 mod composite;
 
 use crate::{rvps::ReferenceValueResolver, token::AttestationTokenBroker};
+pub use challenge::{Challenger, EphemeralJwtChallenger};
 
 use anyhow::{anyhow, Context, Result};
 use canon_json::CanonicalFormatter;
@@ -210,9 +211,9 @@ pub struct VerificationRequest {
 }
 
 pub struct AttestationService {
-    _config: Config,
     rvps: Arc<dyn RvpsApi>,
     token_broker: Box<dyn AttestationTokenBroker + Send + Sync>,
+    challenger: Box<dyn Challenger + Send + Sync>,
 }
 
 /// Transport-neutral runtime status exposed by REST and gRPC AS binaries.
@@ -227,9 +228,21 @@ pub struct ServiceStatus {
 }
 
 impl AttestationService {
-    /// Create a new Attestation Service instance.
+    /// Create a new Attestation Service instance from a parsed [`Config`].
+    ///
+    /// Config-file / CLI entry point, kept for backward compatibility. It
+    /// constructs the RVPS and token-broker *instances* from the config,
+    /// picks a [`Challenger`] (an [`FsJwtChallenger`] at the configured path,
+    /// or the built-in default path when unset), and assembles them via
+    /// [`Self::from_components`]. Pure-lib / wasm consumers that do not want
+    /// to depend on [`Config`] should call [`Self::from_components`] directly
+    /// with their own component instances.
     #[cfg(feature = "fs")]
     pub async fn new(config: Config) -> Result<Self, ServiceError> {
+        use crate::challenge::FsJwtChallenger;
+
+        // Historical `new()` created the work dir at construction time. Kept
+        // as a standalone mkdir purely for behavior parity.
         if !config.work_dir.as_path().exists() {
             fs::create_dir_all(&config.work_dir)
                 .await
@@ -242,11 +255,27 @@ impl AttestationService {
 
         let token_broker = config.attestation_token_broker.to_token_broker()?;
 
-        Ok(Self {
-            _config: config,
+        let challenger: Box<dyn Challenger + Send + Sync> = match config.challenge_key_path {
+            Some(path) => Box::new(FsJwtChallenger::new(path)),
+            None => Box::new(FsJwtChallenger::new(FsJwtChallenger::default_path())),
+        };
+
+        Ok(Self::from_components(rvps, token_broker, challenger))
+    }
+
+    /// Assemble an [`AttestationService`] from already-constructed component
+    /// instances — the fs-free / pure-lib / wasm entry point. No [`Config`],
+    /// no filesystem: supply your own RVPS, token broker, and challenger.
+    pub fn from_components(
+        rvps: Arc<dyn RvpsApi>,
+        token_broker: Box<dyn AttestationTokenBroker + Send + Sync>,
+        challenger: Box<dyn Challenger + Send + Sync>,
+    ) -> Self {
+        Self {
             rvps,
             token_broker,
-        })
+            challenger,
+        }
     }
 
     /// Return AS and verifier dependency status without performing network I/O.
@@ -450,14 +479,10 @@ impl AttestationService {
             .await
     }
 
-    /// Filesystem path of the RSA private key used to sign and verify
-    /// attestation challenge (nonce) tokens. Falls back to the built-in
-    /// default when not set in the config.
-    pub fn challenge_key_path(&self) -> std::path::PathBuf {
-        self._config
-            .challenge_key_path
-            .clone()
-            .unwrap_or_else(challenge::default_challenge_key_path)
+    /// Borrow the underlying [`Challenger`] used to issue/verify challenge
+    /// (nonce) tokens.
+    pub fn challenger(&self) -> &dyn Challenger {
+        self.challenger.as_ref()
     }
 
     pub async fn generate_challenge(
@@ -466,7 +491,7 @@ impl AttestationService {
         tee_parameters: Option<String>,
     ) -> Result<String> {
         match tee {
-            None => challenge::generate_common_challenge(&self.challenge_key_path()),
+            None => self.challenger.generate_challenge().await,
             Some(t) => {
                 self.generate_supplemental_challenge(t, tee_parameters.unwrap_or_default())
                     .await
